@@ -7,57 +7,61 @@
     "build": "webpack",
     "start": "webpack serve",
     "storybook": "storybook dev -p 6006",
-    "lint": "eslint --ext ts,tsx src test"
+    "lint": "eslint",
+    "typecheck": "tsc"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "@carrotsearch/foamtree": "^3.5.1",
-    "@clickhouse/client-web": "^0.2.2",
-    "@skbkontur/icons": "^1.5.0",
+    "@clickhouse/client-web": "^1.10.0",
+    "@skbkontur/icons": "^1.15.2",
     "@skbkontur/react-stack-layout": "^1.2.0",
-    "@skbkontur/react-ui": "^4.15.7",
-    "@types/react": "^18.2.21",
-    "@types/react-dom": "^18.2.7",
-    "classnames": "^2.3.2",
-    "css-loader": "^6.8.1",
+    "@skbkontur/react-ui": "^5.0.8",
+    "@types/react": "^18.3.18",
+    "@types/react-dom": "^18.3.5",
+    "classnames": "^2.5.1",
+    "css-loader": "^7.1.2",
     "fetch-intercept": "^2.4.0",
-    "less-loader": "^11.1.3",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "less-loader": "^12.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-promise-suspense": "^0.3.3",
-    "react-router-dom": "^6.16.0",
-    "style-loader": "^3.3.3",
-    "styled-components": "^6.0.8",
-    "use-debounce": "^10.0.0",
-    "webpack": "^5.88.2",
-    "webpack-cli": "^5.1.4",
-    "whatwg-fetch": "^3.6.19"
+    "react-router-dom": "^6.28.1",
+    "style-loader": "^3.3.4",
+    "styled-components": "^6.1.14",
+    "use-debounce": "^10.0.4",
+    "webpack": "^5.97.1",
+    "webpack-cli": "^6.0.1",
+    "whatwg-fetch": "^3.6.20"
   },
   "devDependencies": {
-    "@babel/cli": "^7.22.15",
-    "@babel/core": "^7.22.17",
+    "@babel/cli": "^7.26.4",
+    "@babel/core": "^7.26.0",
     "@babel/plugin-proposal-class-properties": "^7.18.6",
-    "@babel/preset-env": "^7.22.15",
-    "@babel/preset-react": "^7.22.15",
-    "@babel/preset-typescript": "^7.22.15",
-    "@storybook/addon-essentials": "^7.4.2",
-    "@storybook/addon-interactions": "^7.4.2",
-    "@storybook/addon-links": "^7.4.2",
-    "@storybook/addon-onboarding": "^1.0.8",
-    "@storybook/blocks": "^7.4.2",
-    "@storybook/react": "^7.4.2",
-    "@storybook/react-webpack5": "^7.4.2",
-    "@storybook/testing-library": "^0.2.1",
-    "@typescript-eslint/eslint-plugin": "^6.9.0",
-    "@typescript-eslint/parser": "^6.9.0",
-    "babel-loader": "^9.1.3",
-    "eslint": "^8.52.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-prettier": "^5.0.1",
-    "html-webpack-plugin": "^5.5.3",
-    "storybook": "^7.4.2",
-    "typescript": "^5.2.2",
-    "webpack-dev-server": "^4.15.1"
+    "@babel/preset-env": "^7.26.0",
+    "@babel/preset-react": "^7.26.3",
+    "@babel/preset-typescript": "^7.26.0",
+    "@eslint/js": "^9.18.0",
+    "@storybook/addon-essentials": "^8.4.7",
+    "@storybook/addon-interactions": "^8.4.7",
+    "@storybook/addon-links": "^8.4.7",
+    "@storybook/addon-onboarding": "^8.4.7",
+    "@storybook/blocks": "^8.4.7",
+    "@storybook/react": "^8.4.7",
+    "@storybook/react-webpack5": "^8.4.7",
+    "@storybook/testing-library": "^0.2.2",
+    "@typescript-eslint/eslint-plugin": "^8.20.0",
+    "@typescript-eslint/parser": "^8.20.0",
+    "babel-loader": "^9.2.1",
+    "eslint": "^9.18.0",
+    "eslint-config-prettier": "^10.0.1",
+    "eslint-plugin-prettier": "^5.2.2",
+    "eslint-plugin-storybook": "^0.11.2",
+    "html-webpack-plugin": "^5.6.3",
+    "storybook": "^8.4.7",
+    "typescript": "^5.7.3",
+    "typescript-eslint": "^8.20.0",
+    "webpack-dev-server": "^5.2.0"
   }
 }
