@@ -1,21 +1,21 @@
 import * as React from "react";
-import {ComboBox, MenuSeparator} from "@skbkontur/react-ui";
-import {MediaUiAPlayIcon} from "@skbkontur/icons";
+import { ComboBox, MenuSeparator } from "@skbkontur/react-ui";
+import { MediaUiAPlayIcon } from "@skbkontur/icons";
 
 export const JobComboBox = ({
     value,
     items,
-    handler
+    handler,
 }: {
     value: string | undefined;
     items: string[];
     handler: (item: string | undefined) => void;
 }) => (
-    <ComboBox
+    <ComboBox<string | undefined>
         value={value}
-        getItems={async () => [undefined, <MenuSeparator/>, ...items]}
+        getItems={() => Promise.resolve([undefined, <MenuSeparator />, ...items])}
         onValueChange={x => {
-            if (typeof x === "string" || x == undefined) handler(x);
+            handler(x);
         }}
         itemToValue={x => (typeof x === "string" ? x : "")}
         valueToString={x => (typeof x === "string" ? x : "")}
@@ -23,27 +23,27 @@ export const JobComboBox = ({
         renderValue={x =>
             x == undefined ? (
                 <span>
-                                        {" "}
-                    <MediaUiAPlayIcon/> All jobs
-                                    </span>
+                    {" "}
+                    <MediaUiAPlayIcon /> All jobs
+                </span>
             ) : (
                 <span>
-                                        {" "}
-                    <MediaUiAPlayIcon/> {x}
-                                    </span>
+                    {" "}
+                    <MediaUiAPlayIcon /> {x}
+                </span>
             )
         }
         renderItem={x =>
             x == undefined ? (
                 <span>
-                                        {" "}
-                    <MediaUiAPlayIcon/> All jobs
-                                    </span>
+                    {" "}
+                    <MediaUiAPlayIcon /> All jobs
+                </span>
             ) : (
                 <span>
-                                        {" "}
-                    <MediaUiAPlayIcon/> {x}
-                                    </span>
+                    {" "}
+                    <MediaUiAPlayIcon /> {x}
+                </span>
             )
         }
     />
